requirements.txt
pandas==1.5.3
PyYAML==6.0.1
PyYAML==6.0.1
st_gsheets_connection==0.0.4
streamlit==1.36.0
streamlit_authenticator==0.3.2
~treamlit==1.34.0
~treamlit==1.35.0
